pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "hedvis"
version = "0.1.0"
description = "HED validation, summary, and analysis tools for annotating events and experimental metadata."
readme = "README.md"
authors = [
    { name = "VisLab" },
    { name = "Ian Callanan" },
    { name = "Kay Robbins", email = "Kay.Robbins@utsa.edu" },
]
maintainers = [
    { name = "Kay Robbins", email = "Kay.Robbins@utsa.edu" },
]
license = {text = "MIT"}
keywords = ["HED", "hierarchical event descriptors", "validation", "neuroscience", "BIDS", "events", "metadata", "annotation"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Operating System :: OS Independent",
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
]

requires-python = ">=3.10"

dependencies = [
    "click>=8.0.0",
    "defusedxml",
    "et-xmlfile",
    "inflect",
    "jsonschema",
    "hedtools",
    "matplotlib>=3.9.0",
    "numpy",
    "openpyxl",
    "pandas",
    "portalocker",
    "python-dateutil",
    "pytz",
    "semantic-version",
    "six",
    "wordcloud==1.9.4"
]

[project.urls]
"Homepage" = "https://www.hedtags.org/"
"Documentation" = "https://www.hedtags.org/hed-python/"
"Repository" = "https://github.com/hed-standard/hed-vis/"
"Bug Tracker" = "https://github.com/hed-standard/hed-vis/issues"
"Changelog" = "https://github.com/hed-standard/hed-vis/blob/main/CHANGELOG.md"
"HED Resources" = "https://www.hed-resources.org"

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "codespell>=2.2.0",
    "black>=24.0.0",
    "mdformat>=0.7.0",
    "mdformat-myst>=0.1.5",
]
docs = [
    "sphinx>=7.0.0",
    "furo>=2023.9.10",
    "sphinx-copybutton>=0.5.2",
    "myst-parser>=2.0.0",
    "sphinx-autodoc-typehints>=1.24.0",
    "linkify-it-py>=2.0.0",
]
test = [
    "coverage>=7.0.0",
]



[tool.setuptools.packages.find]
where = [""]
include = ["hedvis*"]
namespaces = false

[tool.setuptools.package-data]
hedvis = []

[tool.codespell]
skip = '*.git,*.pdf,*.svg,versioneer.py,venv*,*.tsv,*.yaml,*.yml,*.json,*.rdf,*.jsonld,spec_tests,,*.xml,*.mediawiki,*.omn,*.toml'
ignore-words-list = 'te,parms,assertIn'

[tool.ruff]
# Target Python 3.10+
target-version = "py310"
line-length = 127

# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
    ".status",
    "status",
]

[tool.ruff.lint]
# Enable flake8-style checks plus additional rules
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "N",     # pep8-naming
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
]

ignore = [
    "E501",   # Line too long (handled by formatter)
    "N802",   # Function name should be lowercase (many test methods use camelCase)
    "N803",   # Argument name should be lowercase (DataFrame df is common)
    "N806",   # Variable in function should be lowercase (common in scientific code)
    # Style rules that are too opinionated
    "SIM108", # if-else-block-instead-of-if-exp (ternary operator preference)
    "SIM105", # suppressible-exception (use contextlib.suppress)
    "SIM118", # in-dict-keys (x in dict vs x in dict.keys())
    "SIM102", # collapsible-if (nested if statements)
    "SIM103", # needless-bool (return bool vs if/else return True/False)
    "SIM110", # reimplemented-builtin (any/all reimplementation)
    "SIM101", # duplicate-isinstance-call
    # Python 3.10+ features
    "UP007",  # Use X | Y for Union types (requires 3.10+)
    "UP008",  # Use super() without parameters (we're okay with explicit)
    "UP031",  # Use format strings over %-formatting (legacy code compatibility)
    "UP028",  # yield in for loop
    "UP045",  # Use X | None for Optional (requires 3.10+)
]

# Allow unused imports in __init__.py files
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.mccabe]
# Maximum complexity (matching your flake8 setting)
max-complexity = 10

[tool.ruff.lint.isort]
known-first-party = ["hedvis"]

[tool.black]
line-length = 127
target-version = ["py310", "py311", "py312", "py313"]
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | __pycache__
  | build
  | dist
  | \.eggs
  | .*\.egg-info
  | \.status
  | status
)/
'''
